% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gnomAD.R
\docType{data}
\name{d_gnomAD}
\alias{d_gnomAD}
\title{GnomAD raw data}
\format{
\subsection{\code{d_gnomAD}}{

A tibble with 30 rows and 2 columns:
\describe{
\item{Position}{mtDNA position.}
\item{Ref}{Reference base.}
\item{Alt}{Variant base.}
\item{HGHomN}{Total from \code{n}.}
\item{N_TLHG}{Number of individuals in the top-level haplogroup.}
\item{n}{Variant counts in different TLHGs.}
\item{TLHG}{TLHGs for the other columns.}
\item{ExcludeReason}{Exclusion reason, \code{NA} if included in \code{d_gnomAD_refined_long}.}
}
}
}
\source{
\url{https://gnomad.broadinstitute.org/news/2020-11-gnomad-v3-1-mitochondrial-dna-variants/}
}
\usage{
d_gnomAD
}
\description{
Only information about homoplasmic variants.
}
\keyword{datasets}
