% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gnomAD.R
\docType{data}
\name{d_gnomAD_refined_long}
\alias{d_gnomAD_refined_long}
\title{GnomAD single nucleotide variation frequency information}
\format{
\subsection{\code{d_gnomAD_refined_long}}{

A tibble with 28,751 rows and 6 columns:
\describe{
\item{Position}{mtDNA position.}
\item{Ref}{Revised Cambridge Reference Sequence (rCRS) base.}
\item{TLHG}{Top-level haplogroup.}
\item{n}{Number of individuals in the TLHG with the \code{Base}.}
\item{Base}{The \code{Base} that \code{n} is a frequency of.}
\item{Type}{The \code{Base} type (either reference or variant).}
\item{N_TLHG}{Number of individuals in the TLHG (note, position specific).}
}
}
}
\source{
\url{https://gnomad.broadinstitute.org/news/2020-11-gnomad-v3-1-mitochondrial-dna-variants/}
}
\usage{
d_gnomAD_refined_long
}
\description{
Information for each single nucleotide variation (SNV) position on
frequencies in each top-level haplogroup (TLHG) for the included positions
(cf. \code{d_gnomAD}). Note, only homoplasmic variants included.
}
\keyword{datasets}
