% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gnomAD.R
\docType{data}
\name{d_gnomAD_TLHG_freq}
\alias{d_gnomAD_TLHG_freq}
\title{GnomAD v. 3.1 TLHG distribution}
\format{
\subsection{\code{d_gnomAD_TLHG_freq}}{

A tibble with 30 rows and 2 columns:
\describe{
\item{TLHG}{Top-level haplogroup.}
\item{N}{Number of individuals in the top-level haplogroup.}
}
}
}
\source{
\url{https://gnomad.broadinstitute.org/news/2020-11-gnomad-v3-1-mitochondrial-dna-variants/}
}
\usage{
d_gnomAD_TLHG_freq
}
\description{
Counts of individuals in top-level haplogroups, averages accross mitogenome
(as a TLHG count is given at each position).
}
\keyword{datasets}
