% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_helix_positioninfo}
\alias{d_helix_positioninfo}
\title{HelixMTdb position information}
\format{
\subsection{\code{d_helix_positioninfo}}{

A tibble with 14,324 rows and 2 columns:
\describe{
\item{Position}{mtDNA position.}
\item{ExcludeReason}{Reason for excluded Position, \code{NA} if not excluded (i.e. included).}
}
}
}
\source{
"A catalog of homoplasmic and heteroplasmic mitochondrial DNA variants in humans" (
\url{https://www.biorxiv.org/content/10.1101/798264v3} / \url{https://doi.org/10.1101/798264})
}
\usage{
d_helix_positioninfo
}
\description{
Information for each position if excluded (why) or not.
}
\keyword{datasets}
