% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helix.R
\docType{data}
\name{d_helix_refined_long}
\alias{d_helix_refined_long}
\title{HelixMTdb single nucleotide variation frequency information}
\format{
\subsection{\code{d_helix_refined_long}}{

A tibble with 28,751 rows and 6 columns:
\describe{
\item{Position}{mtDNA position.}
\item{Ref}{Revised Cambridge Reference Sequence (rCRS) base.}
\item{TLHG}{Top-level haplogroup.}
\item{n}{Number of individuals in the TLHG with the \code{Base}.}
\item{Base}{The \code{Base} that \code{n} is a frequency of.}
\item{Type}{The \code{Base} type (either reference or variant).}
\item{N_TLHG}{Number of individuals in the TLHG, cf. \code{d_helix_TLHG_freq}.}
}
}
}
\source{
"A catalog of homoplasmic and heteroplasmic mitochondrial DNA variants in humans" (
\url{https://www.biorxiv.org/content/10.1101/798264v3} / \url{https://doi.org/10.1101/798264})
}
\usage{
d_helix_refined_long
}
\description{
Information for each single nucleotide variation (SNV) position on
frequencies in each top-level haplogroup (TLHG) for the included positions
(cf. \code{d_helix}). Note, only homoplasmic variants included.
}
\keyword{datasets}
