% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_helix_SNV_freq_long}
\alias{d_helix_SNV_freq_long}
\title{HelixMTdb single nucleotide variation frequency information}
\format{
\subsection{\code{d_helix_SNV_freq_long}}{

A tibble with 28,751 rows and 6 columns:
\describe{
\item{Position}{mtDNA position.}
\item{Ref}{Revised Cambridge Reference Sequence (rCRS) base.}
\item{Alt}{Alternative base.}
\item{TLHG}{Top-level haplogroup}
\item{n_Ref}{Number of individuals in the top-level haplogroup with the reference rCRS allele.}
\item{n_Alt}{Number of individuals in the top-level haplogroup with the alternative allele.}
}
}
}
\source{
"A catalog of homoplasmic and heteroplasmic mitochondrial DNA variants in humans" (
\url{https://www.biorxiv.org/content/10.1101/798264v3} / \url{https://doi.org/10.1101/798264})
}
\usage{
d_helix_SNV_freq_long
}
\description{
Information for each single nucleotide variation (SNV) position on
frequencies in each top-level haplogroup (TLHG) for the included positions
(only binary SNVs included).
}
\keyword{datasets}
