% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helix.R
\docType{data}
\name{d_helix_TLHG_freq}
\alias{d_helix_TLHG_freq}
\title{HelixMTdb TLHG distribution}
\format{
\subsection{\code{d_helix_TLHG_freq}}{

A tibble with 30 rows and 2 columns:
\describe{
\item{TLHG}{Top-level haplogroup.}
\item{N}{Number of individuals in the top-level haplogroup.}
}
}
}
\source{
"A catalog of homoplasmic and heteroplasmic mitochondrial DNA variants in humans" (
\url{https://www.biorxiv.org/content/10.1101/798264v3} / \url{https://doi.org/10.1101/798264})
}
\usage{
d_helix_TLHG_freq
}
\description{
Counts of individuals in top-level haplogroups from HelixMTdb.
}
\keyword{datasets}
