% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-rCRS.R
\docType{data}
\name{d_rCRS}
\alias{d_rCRS}
\title{Revised Cambridge Reference Sequence (rCRS)}
\format{
\subsection{\code{d_rCRS}}{

A tibble with 16,569 rows and 2 columns:
\describe{
\item{Position}{mtDNA position.}
\item{Ref}{Base according to rCRS.}
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/nuccore/NC_012920.1?report=fasta}
}
\usage{
d_rCRS
}
\description{
Revised Cambridge Reference Sequence (rCRS)
}
\examples{
subset(d_rCRS, !(Ref \%in\% c("A", "T", "G", "C")))

}
\keyword{datasets}
