% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_profile.R
\name{extend_profile}
\alias{extend_profile}
\title{Extend profile to all positions with frequency information}
\usage{
extend_profile(
  variants,
  tlhg,
  range = 1:16569,
  d_SNV_long = d_helix_refined_long,
  only_simple_ref_alt = TRUE
)
}
\arguments{
\item{variants}{Vector of variants}

\item{tlhg}{Top-level haplogroup}

\item{range}{Positions range (default = 1:16569)}

\item{d_SNV_long}{Typically, \code{d_helix_refined_long} or \code{d_gnomAD_refined_long}}

\item{only_simple_ref_alt}{Only consider simple (one base) reference and alternative}
}
\description{
Typically, either Helix (\code{d_helix_refined_long}) or
gnomAD (\code{d_gnomAD_refined_long}).
}
\examples{
ext_prof <- extend_profile(c("263G", "16519C"), "H")
ext_prof


}
