% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_profile.R
\name{parse_range}
\alias{parse_range}
\title{Parse range}
\usage{
parse_range(range_str)
}
\arguments{
\item{range_str}{E.g. "1-16569"}
}
\description{
Parse range
}
\examples{
parse_range("1-16569")
parse_range("1-1,2")
parse_range("1-10,222")
parse_range("16024-576")

}
