% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_profile.R
\name{positions_from_variants}
\alias{positions_from_variants}
\title{Get positions from variants}
\usage{
positions_from_variants(x)
}
\arguments{
\item{x}{Vector of variants}
}
\description{
Get positions from variants
}
