% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_profile.R
\name{extend_profile_to_helix_positions}
\alias{extend_profile_to_helix_positions}
\title{Extend profile to all Helix positions}
\usage{
extend_profile_to_helix_positions(variants, tlhg)
}
\arguments{
\item{variants}{Vector of variants}

\item{tlhg}{Top-level haplogroup}
}
\description{
Extend profile to all Helix positions
}
